Fix all known issue about dumping exchanged partition
We have refactored the logic for dumping exchanged partition, all logic for
exchanged partition has moved into the dumping of external/foreign table. That is:

First, dump the partitioned table just like a normal table, there is no logic
for exchanged partition.

Second, dump the external/foreign table, and after the dumping, we will check if
the external table has been exchanged into a partition table, it means that
the TableInfo::parrelid is not zero,  and if so, we will generate a new TocEntry
for EXCHANGE PARTITION statement.

Here, we introduce a new function 'pg_get_partition_path' to construct the access path
for a leaf partition.